build: do not define DESTDIR
DESTDIR is never meant to be defined by a build system, only ever
prepended to paths. Defining it to the value of ${prefix} means that the
prefix is doubled onto the installation path, and in distribution
packaging, where DESTDIR is used, it is overwritten.

The way DESTDIR works is as follows:

./configure --prefix=/usr will install files to /usr/bin /usr/lib
/usr/share etc, and configure the program to look in these locations as
well. Since the configure stage knows how to build software that
contains compiled paths such as sysconfdir, localstatedir, or
datarootdir (/etc, /var, and /usr/share respectively), this must be
internally represented in the build system.

make install DESTDIR="${pkgdir}" or make install DESTDIR=debian/tmp or
similar, is used to tell the software that it will be installed in the
"/usr" prefix, but then actually install it to $DESTDIR/usr

Internally the software *must* still believe it is installed in /usr (as
those are the paths it is configured for), and the only reason DESTDIR
exists is in order to perform virtual installations that install into a
package manager rather than installing onto an actual root filesystem.